+++ src/Main.cs
@@ -294,6 +294,7 @@ public class Initialisation : BaseUnityPlugin
                 Nightlighter.Init();
                 Gadulo.Init();
                 DeathNote.Init();
+                Akelus.Init();
 
                 Lazy.FinalizeGuns(); // Make sure encounter trackables are finalized so shoot styles properly display in the Ammonomicon